更新用户模块

+++ vision-parent/vision-manage/src/main/java/com/vision/service/sys/SysUserService.java
@@ -25,6 +25,6 @@ public interface SysUserService{
 
 	SysUser findUserByName(String userName);
 
-	SysUser findUserByIdWeb(Integer userId);
+	SysUser findUserByIdWeb(Long userId);
 
 }
